requirements.txt
bootstrap4==0.0.5
certifi==2018.11.29
chardet==3.0.4
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.1
django-bootstrap3==10.0.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-registration==2.4.1
django-rest-framework==0.1.0
django-tinymce==2.8.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
Pillow==5.2.0
psycopg2==2.7.5
pyperclip==1.7.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.5
pyuploadcare==2.6.0
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
whitenoise==3.3.1
